refactor(KeyChain): improved readability
