+++ src/types/KeyChain/methods/getPrivateKey.js
@@ -1,0 +1,15 @@
+const {
+  PrivateKey,
+} = require('@dashevo/dashcore-lib');
+
+function getPrivateKey() {
+  let pk;
+  if (this.type === 'HDPrivateKey') {
+    pk = PrivateKey(this.HDPrivateKey.privateKey);
+  }
+  if (this.type === 'privateKey') {
+    pk = PrivateKey(this.privateKey);
+  }
+  return pk;
+}
+module.exports = getPrivateKey;
