+++ src/types/Wallet/methods/fromHDExtPublicKey.js
@@ -1,6 +1,6 @@
 const Dashcore = require('@dashevo/dashcore-lib');
 const { is } = require('../../../utils');
-const KeyChain = require('../../KeyChain');
+const KeyChain = require('../../KeyChain/KeyChain');
 const { WALLET_TYPES } = require('../../../CONSTANTS');
