+++ src/types/Wallet/methods/fromPrivateKey.js
@@ -1,5 +1,5 @@
 const { is } = require('../../../utils');
-const KeyChain = require('../../KeyChain');
+const KeyChain = require('../../KeyChain/KeyChain');
 const { WALLET_TYPES } = require('../../../CONSTANTS');
