+++ src/types/Wallet/methods/fromSeed.js
@@ -1,7 +1,7 @@
 const {
   is,
 } = require('../../../utils');
-const KeyChain = require('../../KeyChain');
+const KeyChain = require('../../KeyChain/KeyChain');
 const { WALLET_TYPES } = require('../../../CONSTANTS');
 
 /**
